csync/pyproject.toml
[project]
name = "csync"
version = "1.0.0"
description = "A beautiful config synchronization tool for managing dotfiles across machines"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "pyyaml>=6.0",
]

[project.scripts]
csync = "csync.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py38"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.isort]
force-single-line = true
known-first-party = ["csync"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"